+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+- none
+
+## [2.3.0] - 2017-01-01
 ### Added
 - Added `allowUseStrict` option to [`no-unused-expression`] to allow `'use strict';` statements.
 - Use [`eslint-ast-utils`] to replace some logic.
@@ -58,7 +61,8 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 - Added [`no-this`] rule
 - Added [`no-throw`] rule
 
-[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.2.0...master
+[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.3.0...master
+[2.3.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.3.0...v2.0.0